+++ lib/lerk.rb
@@ -6,13 +6,12 @@ require 'lerk/lerk'
 require 'lerk/version'
 require 'lerk/excuse'
 require 'lerk/statistics'
+require 'lerk/internal'
+require 'lerk/hive_interface'
 
+require 'lerk/hints'
 require 'lerk/util'
 
 require 'lerk/worker'
 
-require 'lerk/prometheus/util'
-require 'lerk/prometheus/exporter'
-require 'lerk/prometheus/wrapper'
-require 'lerk/prometheus/discord_counter'
-require 'lerk/prometheus/discord_histogram'
+require 'lerk/prometheus'