+++ lib/lerk/hints/local_file_source.rb
@@ -1,0 +1,13 @@
+module Lerk
+  module Hints
+    class LocalFileSource
+      def initialize(path)
+        @path = path
+      end
+
+      def contents
+        IO.read(@path)
+      end
+    end
+  end
+end