+++ lib/lerk/prometheus.rb
@@ -1,0 +1,5 @@
+require 'lerk/prometheus/util'
+require 'lerk/prometheus/exporter'
+require 'lerk/prometheus/wrapper'
+require 'lerk/prometheus/discord_counter'
+require 'lerk/prometheus/discord_histogram'